Refactor code and fix CORS issue
Refactored code into smaller files and addressed CORS errors to enable cross-origin POST requests to external APIs.
